refactor: rename langchain_rag to doc_retriever; retire placeholder responses
The module never used LangChain; it is a deterministic keyword-scan
retriever over DOCS_PATH. Rename it to doc_retriever.py, make docstrings
and docs describe the actual behavior, and retire the placeholder
response paths: extractive answers instead of the hardcoded stub
sentence, no fabricated 'synthetic' citations (module and /query router),
audit.rag_backend now reports keyword_scan. Rewrite docs/rag.md to match
reality and record the decision in ADR 0005. Plan row B.
