+++ tests/test_rag_multi_query.py
@@ -12,7 +12,7 @@ def test_rag_multi_query_and_hyde_flags(tmp_path):
     os.environ["RAG_MULTI_QUERY_ENABLED"] = "true"
     os.environ["RAG_MULTI_QUERY_COUNT"] = "3"
     os.environ["RAG_HYDE_ENABLED"] = "true"
-    # no embeddings provider needed for LC-only stub
+    # no embeddings provider needed for the keyword-scan retriever
     headers = {"X-User-Role": "analyst"}
     r = client.post(
         "/query",
@@ -26,8 +26,8 @@ def test_rag_multi_query_and_hyde_flags(tmp_path):
     body = r.json()
     assert "audit" in body
     audit = body["audit"]
-    # audit backend is langchain now
-    assert audit.get("rag_backend") == "langchain"
+    # audit backend reflects the actual retriever
+    assert audit.get("rag_backend") == "keyword_scan"
     assert audit.get("router_intent") in ("qa", None)
     # flags may not be present in response_audit depending on filtering; check nested audit_dict fallback
     # We at least ensure the call succeeded and returned citations list
